% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pdf.R
\name{compare_pdf}
\alias{compare_pdf}
\title{Generate histogram for one mc variable}
\usage{
compare_pdf(mc_vars, varnames = NULL, fill = "#E00034", unit = "million")
}
\arguments{
\item{mc_vars}{List of Drawings}

\item{varnames}{Names of the variables}

\item{fill}{Colors (RGB) of the inbound fill}
}
\description{
Generate histogram for one mc variable
}
\examples{
sims <- gen_draws(df)
mc_plot(mc_sum(sims))
}
\keyword{histogram}

