% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatMil.R
\name{formatMil}
\alias{formatMil}
\title{Format $ numbers to Millions}
\usage{
formatMil(x, digits = 1)
}
\arguments{
\item{x}{number to format}

\item{digits}{number of digits after comma}
}
\description{
Format $ numbers to Millions
}
\examples{
formatMil(352458967)
}
\keyword{percentiles}

