% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_draws.R
\name{gen_draws}
\alias{gen_draws}
\title{Generate Draws with betapert for the input dataframe given}
\usage{
gen_draws(data, Low = "Low", Medium = "Medium", High = "High",
  Prob = NULL, draws = 50000, bound = 0, refine = F)
}
\arguments{
\item{data}{The input dataframe}

\item{Low}{name of the column for low, same for Medium, High, etc.}

\item{draws}{number of draws defaults to 50000}

\item{bound}{percentile of the low values in the assumption distribution, e.g. 10 means the low is the P10 and High is P90, default is 0}
}
\description{
Generate Draws with betapert for the input dataframe given
}
\examples{
gen_draws(df,Medium = "Nominal")
}
\keyword{dataframe}
\keyword{draws,}
\keyword{multiple,}

