% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_report.R
\name{gen_report}
\alias{gen_report}
\title{Generate Report}
\usage{
gen_report(var, reference)
}
\arguments{
\item{var}{Mc variable}

\item{reference}{reference value for contingency}
}
\description{
Generate Report
}
\examples{
var = rpert(10000,.8,1,1.2)*5000000
gen_report(var)
}
\keyword{statistics}

