% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_hist.R
\name{mc_hist}
\alias{mc_hist}
\title{Generate histogram for one mc variable}
\usage{
mc_hist(mc_var, name = "Variable", fill = "#E00034", unit = "date")
}
\arguments{
\item{name}{Name of the variable}

\item{fill}{Color (RGB) of the inbound fill}

\item{var}{Drawings}
}
\description{
Generate histogram for one mc variable
}
\examples{
sims <- gen_draws(df)
mc_hist(mc_sum(sims))
}
\keyword{histogram}

