% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sum.R
\name{mc_sum}
\alias{mc_sum}
\title{Sum list of variables}
\usage{
mc_sum(list)
}
\arguments{
\item{list}{List of drawings}
}
\description{
Sum list of variables
}
\examples{
sims <- gen_draws(df)
mc_sum(sims)
}
\keyword{drawings}
\keyword{sum,}

