% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_pdf.R
\name{overlay_pdf}
\alias{overlay_pdf}
\title{Generate histogram for one mc variable}
\usage{
overlay_pdf(mc_vars, varnames = NULL, fill = "#E00034", unit = "million",
  alpha0 = 0.8, hide = F)
}
\arguments{
\item{mc_vars}{List of Drawings}

\item{varnames}{Names of the variables}

\item{fill}{Colors (RGB) of the inbound fill}
}
\description{
Generate histogram for one mc variable
}
\examples{
sims <- gen_draws(df)
mc_plot(mc_sum(sims))
}
\keyword{histogram}

