% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_table.R
\name{p_table}
\alias{p_table}
\title{Generate table of percentiles}
\usage{
p_table(x, step = 10, reference = NULL, unit = "million")
}
\arguments{
\item{x}{Drawings}

\item{step}{number of percentiles needed}

\item{reference}{reference cost for contingency}
}
\description{
Generate table of percentiles
}
\examples{
sims <- gen_draws(df)
p_table(mc_sum(sims))
}
\keyword{percentiles}

