% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_table.R
\name{stat_table}
\alias{stat_table}
\title{Generate table of statistics}
\usage{
stat_table(x, reference = NULL, unit = "million", show_odd = F)
}
\arguments{
\item{x}{Drawings}

\item{reference}{reference cost for contingency}
}
\description{
Generate table of statistics
}
\examples{
sims <- gen_draws(df)
stat_table(mc_sum(sims))
}
\keyword{statistics}

