% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_plot.R
\name{tor_plot}
\alias{tor_plot}
\title{Generate tornado chart}
\usage{
tor_plot(df, abrv = F, fill = "#E00034")
}
\arguments{
\item{df}{Tornado dataframe}

\item{abrv}{Abbreviate names or not}

\item{fill}{Color (RGB) of the inbound fill}
}
\description{
Generate tornado chart
}
\examples{
sims <- gen_draws(df)
tor_df = tor_table(sims,mc_sum(sims),top=10)
tor_plot(tor_df)
}
\keyword{chart}
\keyword{tornado,}

