% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_table.R
\name{tor_table}
\alias{tor_table}
\title{Generate table of tornado chart}
\usage{
tor_table(inputs, output, type = "c2v", top = F)
}
\arguments{
\item{inputs}{list of vector of input variables}

\item{output}{vector of output variable}

\item{type}{type of sensitivity analysis (e.g. c2v: contribution to variance, mean impact)}

\item{top}{if integer x is supplied on show top x variables}

\item{x}{Drawings}
}
\description{
Generate table of tornado chart
}
\examples{
sims <- gen_draws(df)
tor_table(sims,mc_sum(sims))
}
\keyword{sensitivity,}
\keyword{tornado}

