% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrisk_rep.R
\name{xrisk_rep}
\alias{xrisk_rep}
\title{Generate Report}
\usage{
xrisk_rep(sims, reference, folder = getwd())
}
\arguments{
\item{reference}{reference value for contingency}

\item{var}{Mc variable}
}
\description{
Generate Report
}
\examples{
var = rpert(10000,.8,1,1.2)*5000000
gen_report(var)
}
\keyword{statistics}

